moved `get_subset_name` logic from pype.plugin to pype.lib
